SVN exceptions: (re-)applied my fix for RAID #278085, because truncated error messages can still be seen in various places, except for commit
The handling of errors with childs was wrong - the line breaks were added at the wrong position with the result of unwanted line breaks and that the last two chars were missing.


Steps:
- perform a checkout (File | Open From Version Control...) for an URL with a lot of files
- press Abort in the Update dialog while the stuff is checked out

expected: shown error message is

Could not save the URL of the version resource
Cancelled by user


actual (prior to this revision): shown error message is

Could not save the URL of the version resource

Cancelled by us



("...by user" vs. "...by us")

git-svn-id: svn://svn.code.sf.net/p/radstudioverins/code/trunk@25 6172efa2-bc72-412b-a7f0-84a214907881